requirements.txt
flask>=3.0.0
paramiko>=3.4.0
pyyaml>=6.0.1
jinja2>=3.1.4
cryptography>=42.0.0
